moved out common react components to js/components
